instrumentation-genai/opentelemetry-instrumentation-langchain/examples/manual/litellm_travel_agent_demo/requirements.txt
deepeval
langchain==1.0.2
langchain-core==1.0.1
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-runtime-inmem==0.14.1
requests
langgraph-supervisor

# OpenTelemetry + GenAI tooling
opentelemetry-api>=1.36.0
opentelemetry-sdk>=1.36.0
opentelemetry-exporter-otlp-proto-grpc>=1.36.0
opentelemetry-instrumentation-requests>=0.57b0
opentelemetry-instrumentation-langchain>=0.36b0

# Add genai utils and evals
splunk-otel-util-genai>=0.1.4
splunk-otel-util-genai-evals>=0.1.4
splunk-otel-genai-evals-deepeval>=0.1.4
splunk-otel-genai-emitters-splunk>=0.1.4